+++ ginga/util/iohelper.py
@@ -10,6 +10,7 @@ import hashlib
 import mimetypes
 import pathlib
 import urllib.parse
+import tempfile
 
 from ginga.misc import Bunch
 
@@ -84,16 +85,33 @@ def guess_filetype(filepath):
     raise ValueError("Can't determine file type of '%s'" % (filepath))
 
 
-def get_fileinfo(filespec, cache_dir='/tmp', download=False):
+def get_download_path(uri, filename, cache_dir):
+
+    # TODO: cache file by uri and timestamp
+    filepath = os.path.join(cache_dir, filename)
+
+    # find a free local copy of the file
+    count, tmpfile = 1, filepath
+    while os.path.exists(tmpfile):
+        pfx, sfx = os.path.splitext(filepath)
+        tmpfile = pfx + '_' + str(count) + sfx
+        count += 1
+    return tmpfile
+
+
+def get_fileinfo(filespec, cache_dir=None):
     """
     Parse a file specification and return information about it.
     """
+    if cache_dir is None:
+        cache_dir = tempfile.gettempdir()
+
     # Loads first science extension by default.
     # This prevents [None] to be loaded instead.
     idx = None
     name_ext = ''
 
-    # User specified an HDU using bracket notation at end of path?
+    # User specified an index using bracket notation at end of path?
     match = re.match(r'^(.+)\[(.+)\]$', filespec)
     if match:
         filespec = match.group(1)
@@ -101,9 +119,13 @@ def get_fileinfo(filespec, cache_dir='/tmp', download=False):
         if ',' in idx:
             hduname, extver = idx.split(',')
             hduname = hduname.strip()
-            extver = int(extver)
-            idx = (hduname, extver)
-            name_ext = "[%s,%d]" % idx
+            if re.match(r'^\d+$', extver):
+                extver = int(extver)
+                idx = (hduname, extver)
+                name_ext = "[%s,%d]" % idx
+            else:
+                idx = idx.strip()
+                name_ext = "[%s]" % idx
         else:
             if re.match(r'^\d+$', idx):
                 idx = int(idx)
@@ -123,18 +145,11 @@ def get_fileinfo(filespec, cache_dir='/tmp', download=False):
         urlinfo = urllib.parse.urlparse(filespec)
         if urlinfo.scheme == 'file':
             # local file
-            filepath = urlinfo.path
-            match = re.match(r"^/(\w+\:)", filepath)
-            if match:
-                # This is a windows path with a drive letter?
-                # strip the leading slash
-                # NOTE: this seems like it should not be necessary and might
-                # break some cases
-                filepath = filepath[1:]
+            filepath = str(pathlib.Path(urlinfo.path))
 
         else:
             path, filename = os.path.split(urlinfo.path)
-            filepath = os.path.join(cache_dir, filename)
+            filepath = get_download_path(url, filename, cache_dir)
 
     else:
         # Not a URL
@@ -148,7 +163,7 @@ def get_fileinfo(filespec, cache_dir='/tmp', download=False):
     name = fname_pfx + name_ext
 
     res = Bunch.Bunch(filepath=filepath, url=url, numhdu=idx,
-                      name=name, ondisk=ondisk)
+                      name=name, idx=name_ext, ondisk=ondisk)
     return res